fix linting & test issues
