+++ src/constants/index.js
@@ -3,3 +3,4 @@
 export * from './button';
 export * from './misc';
 export * from './class';
+export * from './NoncedStyleElement';
