+++ src/constants/misc.js
@@ -1,27 +1,4 @@
 /* @flow */
-/** @jsx node */
-
-/*
-    Alternative jsx-pragmatic/src/component/style.jsx 
-    that forces the use of a nonce prop
-*/
-
-import { Fragment, node, type ChildType, type NullableChildrenType } from 'jsx-pragmatic/src/node';
-
-type StyleProps = {|
-    css : string | {| _getCss : () => string |},
-    nonce? : ?string,
-    children? : ?NullableChildrenType
-|};
-
-export function NoncedStyleElement({ css, nonce, children } : StyleProps) : ChildType {
-    return (
-        <Fragment>
-            <style innerHTML={ typeof css === 'string' ? css : css._getCss() } nonce={ nonce } />
-            { children }
-        </Fragment>
-    );
-}
 
 export const ATTRIBUTE = {
     BUTTON:            ('data-button' : 'data-button'),
