+++ src/ui/buttons/button.jsx
@@ -201,7 +201,7 @@ export function Button({ fundingSource, style, multiple, locale, env, fundingEli
                 <Spinner />
             </div>
 
-            { shouldShowWalletMenu ? <MenuButton textColor={ textColor } nonce={ nonce }/> : null }
+            { shouldShowWalletMenu ? <MenuButton textColor={ textColor } nonce={ nonce } /> : null }
         </div>
     );
 }
