+++ src/ui/tracking.jsx
@@ -2,7 +2,8 @@
 /** @jsx node */
 
 import { node, Fragment, type ChildType } from 'jsx-pragmatic/src';
-import { NoncedStyleElement } from '../constants'
+
+import { NoncedStyleElement } from '../constants';
 
 export function TrackingBeacon({ url, nonce } : {| url : string, nonce : ?string |}) : ChildType {
     return (
